# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AHS Portal API - Allied Health College, Nishtar Medical University
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List documents for an application
      parameters:
        - name: applicationId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Documents list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"

    post:
      operationId: uploadDocument
      tags: [documents]
      summary: Upload a document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocumentBody"
      responses:
        "201":
          description: Document uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{id}:
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # MERIT LISTS
  /merit-lists:
    get:
      operationId: listMeritLists
      tags: [merit]
      summary: List merit lists
      parameters:
        - name: sessionId
          in: query
          schema:
            type: ["integer", "null"]
        - name: programId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Merit lists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MeritList"

    post:
      operationId: createMeritList
      tags: [merit]
      summary: Generate merit list
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMeritListBody"
      responses:
        "201":
          description: Merit list generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeritList"
  /merit-lists/{id}:
    get:
      operationId: getMeritList
      tags: [merit]
      summary: Get merit list with entries
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Merit list with entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeritListWithEntries"

  # VERIFICATION
...
  /notices:
    get:
      operationId: listNotices
      tags: [notices]
      summary: List notices
      parameters:
        - name: active
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Notices list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notice"

    post:
      operationId: createNotice
      tags: [notices]
      summary: Create a notice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoticeBody"
      responses:
        "201":
          description: Notice created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notice"
  /notices/{id}:
    patch:
      operationId: updateNotice
      tags: [notices]
      summary: Update notice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoticeBody"
      responses:
        "200":
          description: Updated notice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notice"

    delete:
      operationId: deleteNotice
      tags: [notices]
      summary: Delete notice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # AUDIT LOGS
...
  /joined-students:
    get:
      operationId: listJoinedStudents
      tags: [students]
      summary: List joined students
      parameters:
        - name: programId
          in: query
          schema:
            type: ["integer", "null"]
        - name: sessionId
          in: query
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Joined students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JoinedStudent"

    post:
      operationId: markStudentJoined
      tags: [students]
      summary: Mark student as joined
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJoinedStudentBody"
      responses:
        "201":
          description: Student marked as joined
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JoinedStudent"
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        fullName:
          type: string
        role:
          type: string
        phone:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, email, fullName, role, isActive, createdAt, updatedAt]
...
    Program:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        duration:
          type: string
        seats:
          type: integer
        description:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, name, code, duration, seats, isActive, createdAt]
...
    AdmissionSession:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        year:
          type: integer
        startDate:
          type: string
        endDate:
          type: string
        isActive:
          type: boolean
        status:
          type: string
        createdAt:
          type: string
      required: [id, name, year, startDate, endDate, isActive, status, createdAt]
...
    Application:
      type: object
      properties:
        id:
          type: integer
        applicationNumber:
          type: string
        userId:
          type: integer
        sessionId:
          type: integer
        programId:
          type: integer
        quotaId:
          type: ["integer", "null"]
        status:
          type: string
        remarks:
          type: ["string", "null"]
        submittedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        user:
          $ref: "#/components/schemas/User"
        program:
          $ref: "#/components/schemas/Program"
        session:
          $ref: "#/components/schemas/AdmissionSession"
      required: [id, applicationNumber, userId, sessionId, programId, status, createdAt, updatedAt, user, program, session]
...
    Document:
      type: object
      properties:
        id:
          type: integer
        applicationId:
          type: integer
        docType:
          type: string
        fileName:
          type: string
        fileUrl:
          type: string
        status:
          type: string
        uploadedAt:
          type: string
      required: [id, applicationId, docType, fileName, fileUrl, status, uploadedAt]
    CreateDocumentBody:
      type: object
      properties:
        applicationId:
          type: integer
        docType:
          type: string
        fileName:
          type: string
        fileUrl:
          type: string
      required: [applicationId, docType, fileName, fileUrl]

    # PAYMENT CHALLANS
...
    MeritList:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        sessionId:
          type: integer
        programId:
          type: integer
        publishedAt:
          type: ["string", "null"]
        isPublished:
          type: boolean
        totalEntries:
          type: integer
        createdAt:
          type: string
      required: [id, name, sessionId, programId, isPublished, totalEntries, createdAt]
    MeritListEntry:
      type: object
      properties:
        id:
          type: integer
        meritListId:
          type: integer
        applicationId:
          type: integer
        rank:
          type: integer
        meritScore:
          type: number
        status:
          type: string
        application:
          $ref: "#/components/schemas/Application"
      required: [id, meritListId, applicationId, rank, meritScore, status, application]
    MeritListWithEntries:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        sessionId:
          type: integer
        programId:
          type: integer
        publishedAt:
          type: ["string", "null"]
        isPublished:
          type: boolean
        totalEntries:
          type: integer
        createdAt:
          type: string
        entries:
          type: array
          items:
            $ref: "#/components/schemas/MeritListEntry"
      required: [id, name, sessionId, programId, isPublished, totalEntries, createdAt, entries]
    CreateMeritListBody:
      type: object
      properties:
        name:
          type: string
        sessionId:
          type: integer
        programId:
          type: integer
      required: [name, sessionId, programId]

    # VERIFICATION
...
    Notice:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        category:
          type: string
        isActive:
          type: boolean
        publishedAt:
          type: ["string", "null"]
        expiresAt:
          type: ["string", "null"]
        createdBy:
          type: integer
        createdAt:
          type: string
      required: [id, title, content, category, isActive, createdBy, createdAt]
    CreateNoticeBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        category:
          type: string
        isActive:
          type: boolean
        expiresAt:
          type: ["string", "null"]
      required: [title, content, category, isActive]

    # AUDIT LOGS
...
    JoinedStudent:
      type: object
      properties:
        id:
          type: integer
        applicationId:
          type: integer
        userId:
          type: integer
        sessionId:
          type: integer
        programId:
          type: integer
        rollNumber:
          type: ["string", "null"]
        joinedAt:
          type: string
        createdAt:
          type: string
        user:
          $ref: "#/components/schemas/User"
        program:
          $ref: "#/components/schemas/Program"
      required: [id, applicationId, userId, sessionId, programId, joinedAt, createdAt, user, program]
    CreateJoinedStudentBody:
      type: object
      properties:
        applicationId:
          type: integer
        rollNumber:
          type: ["string", "null"]
      required: [applicationId]